refactor: move httpjson package to api utils
