+++ server/internal/api/middlewares/middlewares.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"gossip/internal/api"
 	"gossip/internal/repository"
-	"gossip/internal/utils/httpjson"
 	"log/slog"
 	"net/http"
 
@@ -23,7 +22,7 @@ func (m *Middlewares) AuthMiddleware(next http.Handler) http.Handler {
 		sessionIdHeaderValue := r.Header.Get(api.SESSION_ID_HEADER)
 		if sessionIdHeaderValue == "" {
 			slog.Error("empty session ID header")
-			httpjson.WriteError(
+			api.ErrorToJSON(
 				w,
 				http.StatusUnauthorized,
 				invalidSessionIdError,
@@ -33,7 +32,7 @@ func (m *Middlewares) AuthMiddleware(next http.Handler) http.Handler {
 		sessionId, err := uuid.FromString(sessionIdHeaderValue)
 		if err != nil {
 			slog.Error("invalid session ID", "sessionId", sessionId)
-			httpjson.WriteError(
+			api.ErrorToJSON(
 				w,
 				http.StatusUnauthorized,
 				invalidSessionIdError,
@@ -46,7 +45,7 @@ func (m *Middlewares) AuthMiddleware(next http.Handler) http.Handler {
 		)
 		if err != nil {
 			slog.Error("session ID not found", "sessionId", sessionId)
-			httpjson.WriteError(
+			api.ErrorToJSON(
 				w,
 				http.StatusUnauthorized,
 				invalidSessionIdError,
